--- demo-app/src/main/java/org/openmined/syft/demo/MainActivity.kt
@@ -1,50 +1,0 @@
-package org.openmined.syft.demo
-
-import android.os.Bundle
-import androidx.appcompat.app.AppCompatActivity
-import androidx.databinding.DataBindingUtil
-import io.reactivex.Scheduler
-import io.reactivex.android.schedulers.AndroidSchedulers
-import io.reactivex.schedulers.Schedulers
-import kotlinx.android.synthetic.main.activity_main.toolbar
-import org.openmined.syft.demo.databinding.ActivityMainBinding
-import org.openmined.syft.networking.clients.HttpClient
-import org.openmined.syft.networking.clients.SocketClient
-import org.openmined.syft.threading.ProcessSchedulers
-
-class MainActivity : AppCompatActivity() {
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        val binding: ActivityMainBinding =
-                DataBindingUtil.setContentView(this, R.layout.activity_main)
-        setSupportActionBar(toolbar)
-        binding.viewModel = initiateViewModel("192.168.1.13:5000")
-    }
-
-    private fun initiateViewModel(baseUrl: String): FederatedCycleViewModel {
-        val networkingSchedulers = object : ProcessSchedulers {
-            override val computeThreadScheduler: Scheduler
-                get() = Schedulers.io()
-            override val calleeThreadScheduler: Scheduler
-                get() = AndroidSchedulers.mainThread()
-        }
-        val computeSchedulers = object : ProcessSchedulers {
-            override val computeThreadScheduler: Scheduler
-                get() = Schedulers.computation()
-            override val calleeThreadScheduler: Scheduler
-                get() = Schedulers.single()
-        }
-        val socketClient = SocketClient(
-            baseUrl,
-            2000u,
-            computeSchedulers
-        )
-        val httpClient = HttpClient(baseUrl)
-        return FederatedCycleViewModel(
-            socketClient,
-            httpClient,
-            networkingSchedulers,
-            computeSchedulers
-        )
-    }
-}
