+++ demo-app/src/main/java/org/openmined/syft/demo/ui/MainActivity.kt
@@ -1,54 +1,66 @@
 package org.openmined.syft.demo.ui
 
 import android.os.Bundle
-import android.widget.Toast
 import androidx.appcompat.app.AppCompatActivity
-import androidx.lifecycle.Observer
+import androidx.databinding.DataBindingUtil
 import io.reactivex.Scheduler
+import io.reactivex.android.schedulers.AndroidSchedulers
 import io.reactivex.schedulers.Schedulers
-import kotlinx.android.synthetic.main.activity_main.log_area
-import kotlinx.android.synthetic.main.activity_main.start_btn
+import kotlinx.android.synthetic.main.activity_main.toolbar
 import org.openmined.syft.demo.R
+import org.openmined.syft.demo.databinding.ActivityMainBinding
 import org.openmined.syft.demo.datasource.LocalMNISTDataDataSource
 import org.openmined.syft.demo.datasource.LocalMNISTModuleDataSource
 import org.openmined.syft.demo.domain.MNISTDataRepository
 import org.openmined.syft.demo.domain.MNISTModuleRepository
 import org.openmined.syft.demo.domain.MNISTTrainer
+import org.openmined.syft.networking.clients.HttpClient
+import org.openmined.syft.networking.clients.SocketClient
 import org.openmined.syft.threading.ProcessSchedulers
 
+@ExperimentalUnsignedTypes
 class MainActivity : AppCompatActivity() {
-
-    private lateinit var viewModel: MainViewModel
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         setContentView(R.layout.activity_main)
+        val binding: ActivityMainBinding =
+                DataBindingUtil.setContentView(this, R.layout.activity_main)
+        setSupportActionBar(toolbar)
+        binding.viewModel = initiateViewModel("192.168.1.13:5000")
 
-        injectMe()
-        viewModel.trainingState.observe(this, Observer {
-            log_area.append(it.toString())
-        })
-        start_btn.setOnClickListener { viewModel.process().subscribe() }
     }
 
-    private fun injectMe() {
+    private fun initiateViewModel(baseUrl: String): FederatedCycleViewModel {
+        val networkingSchedulers = object : ProcessSchedulers {
+            override val computeThreadScheduler: Scheduler
+                get() = Schedulers.io()
+            override val calleeThreadScheduler: Scheduler
+                get() = AndroidSchedulers.mainThread()
+        }
         val computeSchedulers = object : ProcessSchedulers {
             override val computeThreadScheduler: Scheduler
                 get() = Schedulers.computation()
             override val calleeThreadScheduler: Scheduler
                 get() = Schedulers.single()
         }
+        val socketClient = SocketClient(
+            baseUrl,
+            2000u,
+            computeSchedulers
+        )
+        val httpClient = HttpClient(baseUrl)
 
         val localModuleDataSource = LocalMNISTModuleDataSource(resources, filesDir)
         val moduleRepository = MNISTModuleRepository(localModuleDataSource)
         val localMNISTDataDataSource = LocalMNISTDataDataSource(resources)
         val dataRepository = MNISTDataRepository(localMNISTDataDataSource)
         val trainer = MNISTTrainer()
-        viewModel = MainViewModelFactory(
-            computeSchedulers,
-            moduleRepository,
-            dataRepository,
-            trainer
-        ).create(MainViewModel::class.java)
+        return MainViewModelFactory(
+            socketClient,
+            httpClient,
+            networkingSchedulers,
+            computeSchedulers
+        ).create(FederatedCycleViewModel::class.java)
     }
 }
