+++ demo-app/src/main/java/org/openmined/syft/demo/ui/MainViewModelFactory.kt
@@ -5,21 +5,23 @@ import androidx.lifecycle.ViewModelProvider
 import org.openmined.syft.demo.domain.MNISTModuleRepository
 import org.openmined.syft.demo.domain.MNISTDataRepository
 import org.openmined.syft.demo.domain.MNISTTrainer
+import org.openmined.syft.networking.clients.HttpClient
+import org.openmined.syft.networking.clients.SocketClient
 import org.openmined.syft.threading.ProcessSchedulers
 
 class MainViewModelFactory(
-    private val schedulerProvider: ProcessSchedulers,
-    private val moduleRepository: MNISTModuleRepository,
-    private val dataRepository: MNISTDataRepository,
-    private val trainer: MNISTTrainer
+    private val socketClient: SocketClient,
+    private val httpClient: HttpClient,
+    private val networkSchedulers: ProcessSchedulers,
+    private val computeSchedulers: ProcessSchedulers
 ) : ViewModelProvider.Factory {
 
     override fun <T : ViewModel?> create(modelClass: Class<T>): T {
-        return MainViewModel(
-            schedulerProvider,
-            moduleRepository,
-            dataRepository,
-            trainer
+        return FederatedCycleViewModel(
+            socketClient,
+            httpClient,
+            networkSchedulers,
+            computeSchedulers
         ) as T
     }
 }
