+++ demo-app/src/main/res/layout/activity_main.xml
@@ -7,13 +7,13 @@
 
         <variable
             name="viewModel"
-            type="org.openmined.syft.demo.FederatedCycleViewModel" />
+            type="org.openmined.syft.demo.ui.FederatedCycleViewModel" />
     </data>
 
     <androidx.coordinatorlayout.widget.CoordinatorLayout
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context=".MainActivity">
+        tools:context=".ui.MainActivity">
 
         <com.google.android.material.appbar.AppBarLayout
             android:layout_width="match_parent"
